requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-Mail==0.9.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
pymongo==4.6.1
Werkzeug==3.0.1
flask_cors==4.0.0
gunicorn==20.1.0
pyjwt==2.8.0